backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
sqlalchemy>=2.0.36
pyodbc>=5.0.0
aioodbc>=0.5.0
python-dotenv==1.0.1
Pillow>=10.4.0
requests==2.31.0
pydantic[email]>=2.9.0
scikit-learn>=1.4.0
langchain>=0.1.0
langchain-groq
langchain-community
langchain-huggingface
faiss-cpu
sentence-transformers
